deployments/CCI/Bumper/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
py==1.4.34
pyserial==3.4
pytest==3.2.2
requests==2.18.4
RPi.GPIO==0.6.3
urllib3==1.22
